pyproject.toml
[tool.poetry]
name = "holocene-extinction-game"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-configurations = "^2.2"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13.1"
django-filter = "^2.4.0"
django-extensions = "^3.1.5"
django-import-export = "^2.8.0"
django-storages = "^1.12.3"
Pillow = "^9.1.0"
django-cors-headers = "^3.7.0"
django-querycount = "^0.7.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.2.0"
psycopg2 = "^2.9.3"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.22.1"
boto3 = ">=1.24.14"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
